my minimax

+++ src/rating/Rater.java
@@ -6,5 +6,5 @@ import com.checkers.domain.vo.Field;
  * Created by Eugene on 03.12.2015.
  */
 public interface Rater {
-    int rate(Field field);
+    double rate(Field field);
 }
